new merge fixes
