--- src/test/java/com/sparta/mm/pom/HackerNewsPOMTestsM.java
@@ -1,59 +1,0 @@
-package com.sparta.mm.pom;
-
-import com.sparta.mm.pom.pages.HNHomepage;
-import org.junit.jupiter.api.*;
-import org.openqa.selenium.By;
-import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.support.ui.FluentWait;
-import org.openqa.selenium.support.ui.Wait;
-
-import java.time.Duration;
-import java.util.function.Function;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-
-
-public class HackerNewsPOMTestsM {
-    private static WebDriver driver;
-    private HNHomepage homepage;
-
-    private static final String[] headers = new String[]{"Hacker News", "new", "past", "comments", "ask", "show", "jobs", "submit"};
-
-    @BeforeAll
-    static void setupAll() {
-        System.setProperty("webdriver.chrome.driver", "src/test/resources/chromedriver.exe");
-    }
-
-    @BeforeEach
-    void setup() {
-        driver = new ChromeDriver();
-        homepage = new HNHomepage(driver);
-    }
-
-    @Test
-    @DisplayName("First Score Is Found in 6 Second")
-    void isFoundInTime() {
-        Wait<WebDriver> wait = new FluentWait<>(driver)
-                .withTimeout(Duration.ofSeconds(6))
-                .pollingEvery(Duration.ofSeconds(2))
-                .ignoring(NoSuchElementException.class);
-        driver.get(homepage.goToAsk().getURL());
-        WebElement result = wait.until(driver -> driver.findElement(By.id("score_30940747")));
-        System.out.println(result);
-        Assertions.assertTrue(result.isDisplayed());
-    }
-
-    @AfterEach()
-    void teardown() {
-        driver.close();
-    }
-
-    @AfterAll
-    static void teardownAll() {
-        driver.quit();
-    }
-}
